Move it back to layout.server for now
